Fixed scene example mess using .echo files
