+++ Echo.Core/ext/Scenes/Simple/bunny.obj
