+++ Echo.Core/src/Scenic/ExamplesOld/Sponza.cs
